# src/main/resources/openapi.yaml
openapi: 3.0.3

info:
  title: User API
  description: API REST para creación y gestión de usuarios
  version: 1.0.0

servers:
  - url: http://localhost:8080

security:
  - bearerAuth: []

paths:

  /users:
    post:
      tags: [Users]
      summary: Crear usuario
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            examples:
              ejemplo:
                value:
                  name: "Juan Rodriguez"
                  email: "juan@rodriguez.org"
                  password: "hunter2"
                  phones:
                    - number: "1234567"
                      citycode: "1"
                      contrycode: "57"
      responses:
        '201':
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Users]
      summary: Listar usuarios
      operationId: listUsers
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /users/{id}:
    get:
      tags: [Users]
      summary: Obtener usuario por id
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Users]
      summary: Actualizar usuario completo
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Users]
      summary: Actualizar parcialmente un usuario
      operationId: patchUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            examples:
              ejemplo:
                value:
                  name: "Juan Actualizado"
      responses:
        '200':
          description: Usuario actualizado parcialmente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Users]
      summary: Eliminar usuario
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Usuario eliminado correctamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    UserRequest:
      type: object
      required: [name, email, password, phones]
      properties:
        name:
          type: string
          example: "Juan Rodriguez"
        email:
          type: string
          format: email
          example: "juan@rodriguez.org"
        password:
          type: string
          example: "hunter2"
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneRequest'
    PhoneRequest:
      type: object
      required: [number, citycode, contrycode]
      properties:
        number:
          type: string
          example: "1234567"
        citycode:
          type: string
          example: "1"
        contrycode:
          type: string
          example: "57"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
        token:
          type: string
        isactive:
          type: boolean
        name:
          type: string
        email:
          type: string
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneRequest'
    ErrorResponse:
      type: object
      properties:
        mensaje:
          type: string

  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                mensaje: "Formato de correo inválido"

    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            jwt:
              value:
                mensaje: "Token expirado"

    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                mensaje: "Usuario no encontrado"

    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal:
              value:
                mensaje: "Error interno del servidor"